requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.0
moviepy==1.0.3
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
gdown
psutil

# Google API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email dependencies
emails==0.6.0

beautifulsoup4==4.12.2
lxml==4.9.3 